+++ src/main/java/pavel_kuzenkov/com/github/unique_IP_counter/ip_address_entity/Octet.java
@@ -22,6 +22,6 @@ interface Octet {
      * @return true если адрес уникален(в хранилище ещё нет такого адреса) и был сохран в хранилище,
      * false если хранилище уже содержит такой адрес.
      */
-    boolean addNextOctet(short[] address);
+    boolean addNextOctet(int[] address);
 
 }
